requirements.txt
dj-rest-auth==2.2.5
django-allauth==0.51.0
Django==4.1.1
django-cors-headers==3.13.0
django-filter==23.3
djangorestframework>=3.13.1,!=3.14.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
python-decouple==3.6
djongo==1.3.6
mongoengine==0.28.2
pymongo==3.12.1
